move forward and back more smoother, but to fast

+++ includes/cub3d.h
@@ -6,7 +6,7 @@
 /*   By: jodone <jodone@student.42angouleme.fr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/03/10 10:55:59 by mgarnier          #+#    #+#             */
-/*   Updated: 2026/03/16 16:58:19 by jodone           ###   ########.fr       */
+/*   Updated: 2026/03/17 10:52:02 by jodone           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -82,7 +82,7 @@ int			is_space(char c);
 int			check_map(char **map);
 
 //MOVING
-int			player_move(t_mlx *mlx, int delx, int dely);
+int			player_move(t_mlx *mlx, double delx, double dely);
 
 void		update_frame(void *param);
 void		window_hook(int event, void *par);
